requirements.txt
requests
pyyaml
lxml
wheel